+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Prevent multi select **SelectDropdown** and **Dropdown** from calling `onClick` twice when item is checked ([#1310](https://github.com/optimizely/oui/pull/1310))
 
 ## 45.6.1 - 2020-03-12
 - [Patch] Update Chromatic to be a devDependency ([#1309](https://github.com/optimizely/oui/pull/1309))