+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 45.7.0 - 2020-04-06
 - [Feature] Multiple enhancements to **Select** ([#1313](https://github.com/optimizely/oui/pull/1313))
     - Add `displayError` and `note` to show an error state and message
     - Add `isFullWidth` to make **Select** take 100% width of container